+++ src/utils/reporter.rs
@@ -1,7 +1,8 @@
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 use std::sync::atomic::Ordering;
-use std::sync::mpsc::RecvTimeoutError;
+
+use crossbeam_channel::RecvTimeoutError;
 
 use std::thread;
 use std::thread::Builder;